requirements.txt
# 🔹 Core LangChain ecosystem
langchain==0.3.27
langchain-core==0.3.76
langchain-community==0.3.30
langchain-huggingface==0.3.1
langchain-groq==0.3.8

# 🔹 Vector & traditional retrievers
faiss-cpu==1.12.0
chromadb==1.1.1
scikit-learn==1.7.0          # ✅ Required for TF-IDF retriever
sentence-transformers==5.1.1  # for HuggingFaceEmbeddings

# 🔹 Document loaders & environment handling
pypdf==6.1.0
python-dotenv==1.0.1

# 🔹 Text splitting (needed separately in v0.3+)
langchain-text-splitters==0.3.11

# 🔹 UI / Web App
streamlit==1.50.0

# 🔹 LLM APIs
groq==0.32.0
google-generativeai==0.8.3
